AnimatedTriggerReceiver: use double instead of float for time.
Doing so many incremental subtractions seemed risky
with 32-bit float. Now use 64-bit float for much higher
precision: approx 16 decimal digits compared to 6.

It just makes me feel better about edge cases like
extremely high sample rates.

+++ src/sapphire_vcvrack.hpp
@@ -750,14 +750,14 @@ namespace Sapphire
     };
 
 
-    constexpr float FlashDurationSeconds = 0.05;
+    constexpr double FlashDurationSeconds = 0.05;
 
 
     class AnimatedTriggerReceiver
     {
     private:
         GateTriggerReceiver tr;
-        float flashSecondsRemaining = 0;
+        double flashSecondsRemaining = 0;
 
     public:
         void initialize()
@@ -771,13 +771,13 @@ namespace Sapphire
             return tr.isTriggerActive();
         }
 
-        void update(float voltage, float sampleRateHz)
+        void update(float voltage, double sampleRateHz)
         {
             tr.update(voltage);
             if (tr.isTriggerActive())
                 flashSecondsRemaining = FlashDurationSeconds;
             else if (flashSecondsRemaining > 0)
-                flashSecondsRemaining = std::max<float>(0, flashSecondsRemaining - 1/sampleRateHz);
+                flashSecondsRemaining = std::max<double>(0, flashSecondsRemaining - 1/sampleRateHz);
         }
 
         bool lit() const
